Refactor: Move is_running_with_privileges to common utils
